Add utils package
