+++ frontend/app/src/components/StatusWidget/StatusWidget.tsx
@@ -32,14 +32,13 @@ import {
   BaseButton,
   BaseButtonKind,
   Icon,
-  isNullOrUndefined,
-  notNullOrUndefined,
   Placement,
   ScriptRunState,
   Timer,
   Tooltip,
 } from "@streamlit/lib"
 import { SessionEvent } from "@streamlit/protobuf"
+import { isNullOrUndefined, notNullOrUndefined } from "@streamlit/utils"
 import iconRunning from "@streamlit/app/src/assets/img/icon_running.gif"
 import newYearsRunning from "@streamlit/app/src/assets/img/fireworks.gif"
 import { ConnectionState } from "@streamlit/connection"
