+++ frontend/connection/src/StaticConnection.tsx
@@ -17,10 +17,10 @@
 import { getLogger } from "loglevel"
 
 import { ForwardMsgList } from "@streamlit/protobuf"
+import { localStorageAvailable } from "@streamlit/utils"
 
 import { ConnectionState } from "./ConnectionState"
 import { StreamlitEndpoints } from "./types"
-import { localStorageAvailable } from "./utils"
 
 // TODO: Change this to a stable location and eventually make it configurable
 // Holds url for static asset location
