+++ frontend/connection/src/utils.ts
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+import { isHttps, makePath } from "@streamlit/utils"
+
 import { IS_DEV_ENV, WEBSOCKET_PORT_DEV } from "./constants"
 
 const FINAL_SLASH_RE = /\/+$/
@@ -88,78 +90,4 @@ export function buildWsUri(
   const protocol = isHttps() ? "wss" : "ws"
   const fullPath = makePath(pathname, path)
   return `${protocol}://${hostname}:${port}/${fullPath}`
-}
-
-/**
- * Create an HTTP URI for the given path.
- */
-export function buildHttpUri(
-  { hostname, port, pathname }: URL,
-  path: string
-): string {
-  const protocol = isHttps() ? "https" : "http"
-  const fullPath = makePath(pathname, path)
-  return `${protocol}://${hostname}:${port}/${fullPath}`
-}
-
-export function makePath(basePath: string, subPath: string): string {
-  basePath = basePath.replace(FINAL_SLASH_RE, "").replace(INITIAL_SLASH_RE, "")
-  subPath = subPath.replace(FINAL_SLASH_RE, "").replace(INITIAL_SLASH_RE, "")
-
-  if (basePath.length === 0) {
-    return subPath
-  }
-
-  return `${basePath}/${subPath}`
-}
-
-/**
- * True if we're connected to the host via HTTPS.
- */
-function isHttps(): boolean {
-  return window.location.href.startsWith("https://")
-}
-
-/**
- * Returns cookie value
- */
-export function getCookie(name: string): string | undefined {
-  const r = document.cookie.match(`\\b${name}=([^;]*)\\b`)
-  return r ? r[1] : undefined
-}
-
-// Method taken from
-// https://stackoverflow.com/questions/16427636/check-if-localstorage-is-available
-export function localStorageAvailable(): boolean {
-  const testData = "testData"
-
-  try {
-    const { localStorage } = window
-    localStorage.setItem(testData, testData)
-    localStorage.getItem(testData)
-    localStorage.removeItem(testData)
-  } catch (e) {
-    return false
-  }
-  return true
-}
-
-/**
- * A type predicate that is true if the given value is neither undefined
- * nor null.
- */
-export function notNullOrUndefined<T>(
-  value: T | null | undefined
-): value is T {
-  return <T>value !== null && <T>value !== undefined
-}
-
-/**
- * A type predicate that is true if the given value is either undefined
- * or null.
- */
-export function isNullOrUndefined<T>(
-  value: T | null | undefined
-): value is null | undefined {
-  return <T>value === null || <T>value === undefined
 }
