+++ frontend/lib/src/components/widgets/CustomComponent/ComponentInstance.test.tsx
@@ -28,7 +28,6 @@ import {
   DEFAULT_IFRAME_FEATURE_POLICY,
   DEFAULT_IFRAME_SANDBOX_POLICY,
 } from "~lib/util/IFrameUtil"
-import { buildHttpUri } from "~lib/util/UriUtil"
 import { WidgetStateManager } from "~lib/WidgetStateManager"
 import { bgColorToBaseString, toExportedTheme } from "~lib/theme"
 import { fonts } from "~lib/theme/primitives/typography"
@@ -50,9 +49,15 @@ import { ComponentMessageType, StreamlitMessageType } from "./enums"
 vi.useFakeTimers()
 
 // Mock uri utils.
-vi.mock("~lib/util/UriUtil")
-const mockedBuildHttpUri = buildHttpUri as Mock
-mockedBuildHttpUri.mockImplementation(() => "registry/url")
+vi.mock("@streamlit/utils", async () => {
+  const actualModule = await vi.importActual("@streamlit/utils")
+  const mockedBuildHttpUri = vi.fn().mockImplementation(() => "registry/url")
+
+  return {
+    ...actualModule,
+    buildHttpUri: mockedBuildHttpUri,
+  }
+})
 
 // Mock our WidgetStateManager
 vi.mock("~lib/WidgetStateManager")
