+++ frontend/lib/src/util/utils.test.ts
@@ -19,7 +19,6 @@ import { MockInstance } from "vitest"
 import {
   EMBED_QUERY_PARAM_KEY,
   EMBED_QUERY_PARAM_VALUES,
-  getCookie,
   getEmbedUrlParams,
   getLoadingScreenType,
   isColoredLineDisplayed,
@@ -34,54 +33,6 @@ import {
   preserveEmbedQueryParams,
   setCookie,
 } from "./utils"
-
-describe("getCookie", () => {
-  afterEach(() => {
-    document.cookie.split(";").forEach(cookie => {
-      const eqPos = cookie.indexOf("=")
-      const name = eqPos > -1 ? cookie.substr(0, eqPos) : cookie
-      document.cookie = `${name}=;expires=Thu, 01 Jan 1970 00:00:00 GMT`
-    })
-  })
-
-  it("get existing cookie", () => {
-    document.cookie = "flavor=chocolatechip"
-    const cookie = getCookie("flavor")
-    expect(cookie).toEqual("chocolatechip")
-  })
-
-  it("get missing cookie", () => {
-    document.cookie = "sweetness=medium;"
-    document.cookie = "flavor=chocolatechip;"
-    document.cookie = "type=darkchocolate;"
-    const cookie = getCookie("recipe")
-    expect(cookie).toEqual(undefined)
-  })
-
-  it("find cookie in the front", () => {
-    document.cookie = "flavor=chocolatechip;"
-    document.cookie = "sweetness=medium;"
-    document.cookie = "type=darkchocolate;"
-    const cookie = getCookie("flavor")
-    expect(cookie).toEqual("chocolatechip")
-  })
-
-  it("find cookie in the middle", () => {
-    document.cookie = "sweetness=medium;"
-    document.cookie = "flavor=chocolatechip;"
-    document.cookie = "type=darkchocolate;"
-    const cookie = getCookie("flavor")
-    expect(cookie).toEqual("chocolatechip")
-  })
-
-  it("find cookie in the end", () => {
-    document.cookie = "sweetness=medium;"
-    document.cookie = "type=darkchocolate;"
-    document.cookie = "flavor=chocolatechip;"
-    const cookie = getCookie("flavor")
-    expect(cookie).toEqual("chocolatechip")
-  })
-})
 
 describe("setCookie", () => {
   afterEach(() => {
