+++ frontend/utils/package.json
@@ -1,18 +1,18 @@
 {
-  "name": "@streamlit/connection",
+  "name": "@streamlit/utils",
   "version": "1.41.1",
   "license": "Apache-2.0",
   "private": true,
   "type": "module",
-  "main": "./dist/streamlit-connection.cjs.js",
-  "module": "./dist/streamlit-connection.es.js",
-  "browser": "./dist/streamlit-connection.umd.js",
+  "main": "./dist/streamlit-utils.cjs.js",
+  "module": "./dist/streamlit-utils.es.js",
+  "browser": "./dist/streamlit-utils.umd.js",
   "types": "./dist/index.d.ts",
   "exports": {
     ".": {
-      "require": "./dist/streamlit-connection.cjs.js",
-      "import": "./dist/streamlit-connection.es.js",
-      "default": "./dist/streamlit-connection.umd.js"
+      "require": "./dist/streamlit-utils.cjs.js",
+      "import": "./dist/streamlit-utils.es.js",
+      "default": "./dist/streamlit-utils.umd.js"
     }
   },
   "scripts": {
@@ -26,22 +26,15 @@
     "formatCheck": "prettier --check --config ../.prettierrc --ignore-path ../.prettierignore './src/**/*.{js,ts,jsx,tsx}'"
   },
   "dependencies": {
-    "@streamlit/protobuf": "workspace:^",
-    "@streamlit/utils": "workspace:^",
-    "axios": "^1.7.9",
-    "lodash": "^4.17.21",
-    "loglevel": "^1.9.2"
+    "lodash": "^4.17.21"
   },
   "devDependencies": {
-    "axios-mock-adapter": "^1.21.2",
     "eslint": "^8.33.0",
     "prettier": "^2.8.3",
     "vite": "^5.4.12",
     "vite-plugin-dts": "^4.5.0",
-    "vite-plugin-svgr": "^4.2.0",
     "vite-tsconfig-paths": "^4.3.1",
-    "vitest": "^2.1.2",
-    "vitest-websocket-mock": "^0.3.0"
+    "vitest": "^2.1.2"
   },
   "browserslist": [
     ">0.2%",
