+++ frontend/utils/src/types/index.ts
@@ -14,14 +14,21 @@
  * limitations under the License.
  */
 
-// Note: Cached themes before version 1 were simply stored with key equal to
-// CACHED_THEME_BASE_KEY (with no version number).
-const CACHED_THEME_VERSION = 1
-const CACHED_THEME_BASE_KEY = `stActiveTheme-${window.location.pathname}`
-
-export const LocalStore = {
-  CACHED_THEME_VERSION,
-  CACHED_THEME_BASE_KEY,
-
-  ACTIVE_THEME: `${CACHED_THEME_BASE_KEY}-v${CACHED_THEME_VERSION}`,
+/**
+ * A type predicate that is true if the given value is neither undefined
+ * nor null.
+ */
+export function notNullOrUndefined<T>(
+  value: T | null | undefined
+): value is T {
+  return <T>value !== null && <T>value !== undefined
+}
+/**
+ * A type predicate that is true if the given value is either undefined
+ * or null.
+ */
+export function isNullOrUndefined<T>(
+  value: T | null | undefined
+): value is null | undefined {
+  return <T>value === null || <T>value === undefined
 }
