+++ frontend/utils/src/uri/index.test.ts
@@ -1,0 +1,89 @@
+/**
+ * Copyright (c) Streamlit Inc. (2018-2022) Snowflake Inc. (2022-2025)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { buildHttpUri, isHttps, makePath } from "."
+
+const location: Partial<Location> = {}
+
+global.window = Object.create(window)
+Object.defineProperty(window, "location", { value: location })
+
+describe("uri", () => {
+  describe("buildHttpUri", () => {
+    it("builds HTTP URI correctly", () => {
+      location.href = "http://something"
+      const uri = buildHttpUri(
+        {
+          hostname: "the_host",
+          port: "9988",
+          pathname: "foo/bar",
+        } as URL,
+        "baz"
+      )
+      expect(uri).toBe("http://the_host:9988/foo/bar/baz")
+    })
+
+    it("builds HTTPS URI correctly", () => {
+      location.href = "https://something"
+      const uri = buildHttpUri(
+        {
+          hostname: "the_host",
+          port: "9988",
+          pathname: "foo/bar",
+        } as URL,
+        "baz"
+      )
+      expect(uri).toBe("https://the_host:9988/foo/bar/baz")
+    })
+
+    it("builds HTTP URI with no base path", () => {
+      location.href = "http://something"
+      const uri = buildHttpUri(
+        {
+          hostname: "the_host",
+          port: "9988",
+          pathname: "",
+        } as URL,
+        "baz"
+      )
+      expect(uri).toBe("http://the_host:9988/baz")
+    })
+  })
+
+  describe("makePath", () => {
+    it("makes path correctly", () => {
+      const path = makePath("foo/bar", "baz")
+      expect(path).toBe("foo/bar/baz")
+    })
+
+    it("makes path with no base path", () => {
+      const path = makePath("", "baz")
+      expect(path).toBe("baz")
+    })
+  })
+
+  describe("isHttps", () => {
+    it("returns true for HTTPS", () => {
+      location.href = "https://something"
+      expect(isHttps()).toBe(true)
+    })
+
+    it("returns false for HTTP", () => {
+      location.href = "http://something"
+      expect(isHttps()).toBe(false)
+    })
+  })
+})
