+++ frontend/utils/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "root": true,
+  "extends": "@streamlit/typescript-config/base.tsconfig.json",
+  "include": ["src/**/*"]
+}
